+++ Optimizer/EdgeForm.Designer.cs
@@ -28,7 +28,6 @@
         /// </summary>
         private void InitializeComponent()
         {
-            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(EdgeForm));
             this.label1 = new System.Windows.Forms.Label();
             this.label2 = new System.Windows.Forms.Label();
             this.textBox1 = new System.Windows.Forms.TextBox();
@@ -41,9 +40,10 @@
             this.label1.AutoSize = true;
             this.label1.Font = new System.Drawing.Font("Segoe UI Semibold", 15F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.label1.ForeColor = System.Drawing.Color.DodgerBlue;
-            this.label1.Location = new System.Drawing.Point(12, 9);
+            this.label1.Location = new System.Drawing.Point(10, 7);
+            this.label1.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(426, 35);
+            this.label1.Size = new System.Drawing.Size(341, 28);
             this.label1.TabIndex = 0;
             this.label1.Tag = "themeable";
             this.label1.Text = "Enter your desired download folder:";
@@ -51,14 +51,15 @@
             // label2
             // 
             this.label2.AutoSize = true;
-            this.label2.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.label2.ForeColor = System.Drawing.Color.DodgerBlue;
-            this.label2.Location = new System.Drawing.Point(13, 84);
+            this.label2.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, ((System.Drawing.FontStyle)((System.Drawing.FontStyle.Bold | System.Drawing.FontStyle.Italic))), System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.label2.ForeColor = System.Drawing.Color.Silver;
+            this.label2.Location = new System.Drawing.Point(10, 67);
+            this.label2.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.label2.Name = "label2";
-            this.label2.Size = new System.Drawing.Size(124, 28);
+            this.label2.Size = new System.Drawing.Size(110, 21);
             this.label2.TabIndex = 34;
-            this.label2.Tag = "themeable";
-            this.label2.Text = "Example: X:\\";
+            this.label2.Tag = "";
+            this.label2.Text = "(Example: X:\\)";
             // 
             // textBox1
             // 
@@ -66,9 +67,10 @@
             this.textBox1.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.textBox1.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.textBox1.ForeColor = System.Drawing.Color.White;
-            this.textBox1.Location = new System.Drawing.Point(18, 47);
+            this.textBox1.Location = new System.Drawing.Point(14, 38);
+            this.textBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.textBox1.Name = "textBox1";
-            this.textBox1.Size = new System.Drawing.Size(472, 34);
+            this.textBox1.Size = new System.Drawing.Size(378, 29);
             this.textBox1.TabIndex = 0;
             // 
             // cancel
@@ -81,9 +83,10 @@
             this.cancel.FlatAppearance.MouseOverBackColor = System.Drawing.Color.RoyalBlue;
             this.cancel.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
             this.cancel.ForeColor = System.Drawing.Color.White;
-            this.cancel.Location = new System.Drawing.Point(306, 135);
+            this.cancel.Location = new System.Drawing.Point(245, 108);
+            this.cancel.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.cancel.Name = "cancel";
-            this.cancel.Size = new System.Drawing.Size(89, 39);
+            this.cancel.Size = new System.Drawing.Size(71, 31);
             this.cancel.TabIndex = 36;
             this.cancel.Tag = "themeable";
             this.cancel.Text = "Cancel";
@@ -99,23 +102,24 @@
             this.ok.FlatAppearance.MouseOverBackColor = System.Drawing.Color.RoyalBlue;
             this.ok.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
             this.ok.ForeColor = System.Drawing.Color.White;
-            this.ok.Location = new System.Drawing.Point(401, 135);
+            this.ok.Location = new System.Drawing.Point(321, 108);
+            this.ok.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.ok.Name = "ok";
-            this.ok.Size = new System.Drawing.Size(89, 39);
+            this.ok.Size = new System.Drawing.Size(71, 31);
             this.ok.TabIndex = 35;
             this.ok.Tag = "themeable";
             this.ok.Text = "OK";
             this.ok.UseVisualStyleBackColor = false;
             this.ok.Click += new System.EventHandler(this.ok_Click);
             // 
-            // EdgeFolder
+            // EdgeForm
             // 
             this.AcceptButton = this.ok;
-            this.AutoScaleDimensions = new System.Drawing.SizeF(120F, 120F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(96F, 96F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi;
             this.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(20)))), ((int)(((byte)(20)))));
             this.CancelButton = this.cancel;
-            this.ClientSize = new System.Drawing.Size(502, 186);
+            this.ClientSize = new System.Drawing.Size(402, 149);
             this.Controls.Add(this.cancel);
             this.Controls.Add(this.ok);
             this.Controls.Add(this.textBox1);
@@ -125,10 +129,10 @@
             this.Font = new System.Drawing.Font("Segoe UI Semibold", 9F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.ForeColor = System.Drawing.Color.White;
             this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
-            this.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
+            this.Margin = new System.Windows.Forms.Padding(2, 3, 2, 3);
             this.MaximizeBox = false;
             this.MinimizeBox = false;
-            this.Name = "EdgeFolder";
+            this.Name = "EdgeForm";
             this.ShowIcon = false;
             this.ShowInTaskbar = false;
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;