+++ Optimizer/SilentConfig.cs
@@ -46,6 +46,7 @@ namespace Optimizer
         public bool? DisableFeatureUpdates { get; set; }
         public bool? EnableLongPaths { get; set; }
         public bool? RemoveCastToDevice { get; set; }
+        public bool? DisableActionCenter { get; set; }
 
         public bool? DisableOneDrive { get; set; }
     }